Blood Bank
A simple blood bank system where donors can fill in their details (name, age, gender, contact details, address, blood group). Their details can be viewed or deleted.

For the benefit of the blood bank, the project also has provision to count the number of donors for a particular blood group and search for details of a particular donor.

The details of the donors are stored in binary files.